Bug with timeout when connecting through ssl fixed
PHP has a bug when connecting to imap through ssl, ha timeout
connection cannot be set. Due to this, when a connection cannot
be stablished, the process is blocked more or less 20 minutes.
To avoid this, we have created an optional "sslSecureConnection",
(disabled by default) which makes a fork of the process that makes
the connecftion and the parent is in charge of killing the son if
the connection time takes more than X seconds (by default 10)